+++ app/controllers/api/v1/rubygems_controller.rb
@@ -1,11 +1,9 @@
 class Api::V1::RubygemsController < Api::BaseController
-  skip_before_action :verify_authenticity_token, :only => [:create, :yank, :unyank]
+  skip_before_action :verify_authenticity_token, :only => [:create]
 
-  before_action :authenticate_with_api_key, :only => [:index, :create, :yank, :unyank]
-  before_action :verify_authenticated_user, :only => [:index, :create, :yank, :unyank]
+  before_action :authenticate_with_api_key, :only => [:index, :create]
+  before_action :verify_authenticated_user, :only => [:index, :create]
   before_action :find_rubygem,              :only => [:show]
-  before_action :find_rubygem_by_name,      :only => [:yank, :unyank]
-  before_action :validate_gem_and_version,  :only => [:yank]
 
   def index
     @rubygems = current_user.rubygems.with_versions
@@ -32,45 +30,12 @@ class Api::V1::RubygemsController < Api::BaseController
     render :text => gemcutter.message, :status => gemcutter.code
   end
 
-  def yank
-    if @version.indexed?
-      @version.yank!
-      current_user.deletions.create! rubygem: @version.rubygem.name, number: @version.number, platform: @version.platform
-      StatsD.increment 'yank.success'
-      render :text => "Successfully yanked gem: #{@version.to_title}"
-    else
-      StatsD.increment 'yank.failure'
-      render :text => "The version #{params[:version]} has already been yanked.", :status => :unprocessable_entity
-    end
-  end
-
-  def unyank
-    render text: "Unyanking of gems is no longer supported.", status: :gone
-  end
-
   def reverse_dependencies
     names = Rubygem.reverse_dependencies(params[:id]).pluck(:name)
 
     respond_to do |format|
       format.json { render json: names }
       format.yaml { render yaml: names, yamlish: true }
-    end
-  end
-
-  private
-
-  def validate_gem_and_version
-    if !@rubygem.hosted?
-      render :text => "This gem does not exist.", :status => :not_found
-    elsif !@rubygem.owned_by?(current_user)
-      render :text => "You do not have permission to yank this gem.", :status => :forbidden
-    else
-      begin
-        slug = params[:platform].blank? ? params[:version] : "#{params[:version]}-#{params[:platform]}"
-        @version = Version.find_from_slug!(@rubygem, slug)
-      rescue ActiveRecord::RecordNotFound
-        render :text => "The version #{params[:version]} does not exist.", :status => :not_found
-      end
     end
   end
 end
