+++ app/controllers/application_controller.rb
@@ -56,16 +56,6 @@ class ApplicationController < ActionController::Base
     end
   end
 
-  def find_rubygem_by_name
-    @url      = params[:url]
-    @gem_name = params[:gem_name]
-    @rubygem  = Rubygem.find_by_name(@gem_name)
-    if @rubygem.nil? && @gem_name != WebHook::GLOBAL_PATTERN
-      render :text   => "This gem could not be found",
-             :status => :not_found
-    end
-  end
-
   def set_page
     @page = [1, params[:page].to_i].max
   end
